backend/requirements.txt
# backend/requirements.txt
flask
flask-cors
python-dotenv
spotipy
scikit-learn
numpy
pandas
gunicorn
requests
python-jose
